Extract shared services for members and people (#860)
Move app.routers.{members,people} and their app.mcp.admin counterparts'
near-identical CRUD/merge bodies into
app/services/{members,people}_service.py. Both keep the HTTPException
convention (matching the pre-existing shared helpers they consolidate).

app.routers.registrations picks up its parse_phone import from
people_service now that it no longer lives in app.routers.people.
